+++ src/Modèle/Key.java
@@ -1,7 +1,7 @@
 package Modèle;
 
 public class Key extends Objet{
-    public Key(String nom, int posLargeur, int posHauteur) {
+    public Key(String nom, double posLargeur, double posHauteur) {
         super(nom, posLargeur, posHauteur);
     }
 }
